+++ tests/testthat/test-fit_interfaces.R
@@ -156,6 +156,7 @@ test_that("overhead of parsnip interface is minimal (#1071)", {
   skip_on_cran()
   skip_on_covr()
   skip_if_not_installed("bench")
+  skip_if_not_installed("parsnip", minimum_version = "1.3.0")
 
   bm <- bench::mark(
     time_engine = lm(mpg ~ ., mtcars),
@@ -167,10 +168,10 @@ test_that("overhead of parsnip interface is minimal (#1071)", {
 
   expect_true(
     bm$median[2] < 3.5,
-    label = paste0("parsnip overhead factor (formula interface): ", bm$median[2])
+    label = paste0("parsnip overhead factor (formula interface): ", round(bm$median[2], 4))
   )
   expect_true(
     bm$median[3] < 3.75,
-    label = paste0("parsnip overhead factor (xy interface): ", bm$median[3])
+    label = paste0("parsnip overhead factor (xy interface): ", round(bm$median[3], 4))
   )
 })