+++ tests/testthat/test-linear_reg.R
@@ -358,3 +358,32 @@ test_that("check_args() works", {
     }
   )
 })
+
+
+test_that("prevent using a Poisson family", {
+  skip_if(rlang::is_installed("glmnet"))
+  expect_snapshot(
+    linear_reg(penalty = 1) %>%
+      set_engine("glmnet", family = poisson) %>%
+      fit(mpg ~ ., data = mtcars),
+    error = TRUE
+  )
+  expect_snapshot(
+    linear_reg(penalty = 1) %>%
+      set_engine("glmnet", family = stats::poisson) %>%
+      fit(mpg ~ ., data = mtcars),
+    error = TRUE
+  )
+  expect_snapshot(
+    linear_reg(penalty = 1) %>%
+      set_engine("glmnet", family = stats::poisson()) %>%
+      fit(mpg ~ ., data = mtcars),
+    error = TRUE
+  )
+  expect_snapshot(
+    linear_reg(penalty = 1) %>%
+      set_engine("glmnet", family = "poisson") %>%
+      fit(mpg ~ ., data = mtcars),
+    error = TRUE
+  )
+})